requirements.txt
mysql_connector_python==9.0.0
qrcode==7.4.2
rich==13.9.2